# Speed up boot time by adding default spec files to stub lines