--- a/lib/rubygems/specification.rb
+++ b/lib/rubygems/specification.rb
@@ -2462,14 +2462,14 @@ class Gem::Specification < Gem::BasicSpecification
   # be eval'ed and reconstruct the same specification later.  Attributes that
   # still have their default values are omitted.
 
-  def to_ruby(include_stub_files: true)
+  def to_ruby
     mark_version
     result = []
     result << "# -*- encoding: utf-8 -*-"
     result << "#{Gem::StubSpecification::PREFIX}#{name} #{version} #{platform} #{raw_require_paths.join("\0")}"
     result << "#{Gem::StubSpecification::PREFIX}#{extensions.join "\0"}" unless
       extensions.empty?
-    result << "#{Gem::StubSpecification::FILES_PREFIX}#{files.join("\0")}" if include_stub_files && default_gem?
+    result << "#{Gem::StubSpecification::FILES_PREFIX}#{files.join("\0")}" if default_gem?
     result << nil
     result << "Gem::Specification.new do |s|"