--- a/lib/rubygems/stub_specification.rb
+++ b/lib/rubygems/stub_specification.rb
@@ -35,11 +35,8 @@ class Gem::StubSpecification < Gem::BasicSpecification
     def initialize(data, extensions)
       parts          = data[PREFIX.length..-1].split(" ".freeze, 4)
       @name          = parts[0].freeze
-      @version       = if Gem::Version.correct?(parts[1])
-        Gem::Version.new(parts[1])
-      else
-        Gem::Version.new(0)
-      end
+      parts.insert(1, 0) unless Gem::Version.correct?(parts[1])
+      @version       = Gem::Version.new(parts[1])
 
       @platform      = Gem::Platform.new parts[2]
       @extensions    = extensions