## What was the end-user or developer problem that led to this PR?

Rubygems boot time is slow. One of the causes if that all default gems are evaluated, in order to figure out which files they include and be able to "upgrade requires" in case a higher version is installed. So for example, when `require "psych"` is run, rubygems checks whether there's any file named "psych.rb` inside a default gem, and in that case activates the highest version of that default gem in the system.

This works, but evaluating all default specs in order to do this is expensive.

## What is your fix for the problem, implemented in this PR?

My fix is, similarly to what we do with other information inside gemspecs to avoid having to actually re-evaluate once they are insalled, is to create a new stub line including the list of files inside the gem. So, before this patch, the `psych` gemspec installed would like like this:

```ruby
# -*- encoding: utf-8 -*-
# stub: psych 3.3.0 ruby lib
# stub: ext/psych/extconf.rb

Gem::Specification.new do |s|
  s.name = "psych".freeze
  s.version = "3.3.0"

  s.required_rubygems_version = Gem::Requirement.new(">= 0".freeze) if s.respond_to? :required_rubygems_version=
  s.require_paths = ["lib".freeze]
  s.authors = ["Aaron Patterson".freeze, "SHIBATA Hiroshi".freeze, "Charles Oliver Nutter".freeze]
  s.date = "2021-07-08"
  s.description = "Psych is a YAML parser and emitter. Psych leverages libyaml[https://pyyaml.org/wiki/LibYAML]\nfor its YAML parsing and emitting capabilities. In addition to wrapping libyaml,\nPsych also knows how to serialize and de-serialize most Ruby objects to and from the YAML format.\n".freeze
  s.email = ["aaron@tenderlovemaking.com".freeze, "hsbt@ruby-lang.org".freeze, "headius@headius.com".freeze]
  s.extensions = ["ext/psych/extconf.rb".freeze]
  s.extra_rdoc_files = ["README.md".freeze]
  s.files = ["README.md".freeze, "ext/psych/extconf.rb".freeze, "psych.rb".freeze, "psych.so".freeze, "psych/class_loader.rb".freeze, "psych/coder.rb".freeze, "psych/core_ext.rb".freeze, "psych/exception.rb".freeze, "psych/handler.rb".freeze, "psych/handlers/document_stream.rb".freeze, "psych/handlers/recorder.rb".freeze, "psych/json/ruby_events.rb".freeze, "psych/json/stream.rb".freeze, "psych/json/tree_builder.rb".freeze, "psych/json/yaml_events.rb".freeze, "psych/nodes.rb".freeze, "psych/nodes/alias.rb".freeze, "psych/nodes/document.rb".freeze, "psych/nodes/mapping.rb".freeze, "psych/nodes/node.rb".freeze, "psych/nodes/scalar.rb".freeze, "psych/nodes/sequence.rb".freeze, "psych/nodes/stream.rb".freeze, "psych/omap.rb".freeze, "psych/parser.rb".freeze, "psych/scalar_scanner.rb".freeze, "psych/set.rb".freeze, "psych/stream.rb".freeze, "psych/streaming.rb".freeze, "psych/syntax_error.rb".freeze, "psych/tree_builder.rb".freeze, "psych/versions.rb".freeze, "psych/visitors.rb".freeze, "psych/visitors/depth_first.rb".freeze, "psych/visitors/emitter.rb".freeze, "psych/visitors/json_tree.rb".freeze, "psych/visitors/to_ruby.rb".freeze, "psych/visitors/visitor.rb".freeze, "psych/visitors/yaml_tree.rb".freeze, "psych/y.rb".freeze]
  s.homepage = "https://github.com/ruby/psych".freeze
  s.licenses = ["MIT".freeze]
  s.rdoc_options = ["--main".freeze, "README.md".freeze]
  s.required_ruby_version = Gem::Requirement.new(">= 2.4.0".freeze)
  s.rubygems_version = "3.3.0.dev".freeze
  s.summary = "Psych is a YAML parser and emitter".freeze
end
```

but now it looks like this (note the 4th comment on top of the file):

```ruby
# -*- encoding: utf-8 -*-
# stub: psych 3.3.0 ruby lib
# stub: ext/psych/extconf.rb
# files-stub: README.mdext/psych/extconf.rbpsych.rbpsych.sopsych/class_loader.rbpsych/coder.rbpsych/core_ext.rbpsych/exception.rbpsych/handler.rbpsych/handlers/document_stream.rbpsych/handlers/recorder.rbpsych/json/ruby_events.rbpsych/json/stream.rbpsych/json/tree_builder.rbpsych/json/yaml_events.rbpsych/nodes.rbpsych/nodes/alias.rbpsych/nodes/document.rbpsych/nodes/mapping.rbpsych/nodes/node.rbpsych/nodes/scalar.rbpsych/nodes/sequence.rbpsych/nodes/stream.rbpsych/omap.rbpsych/parser.rbpsych/scalar_scanner.rbpsych/set.rbpsych/stream.rbpsych/streaming.rbpsych/syntax_error.rbpsych/tree_builder.rbpsych/versions.rbpsych/visitors.rbpsych/visitors/depth_first.rbpsych/visitors/emitter.rbpsych/visitors/json_tree.rbpsych/visitors/to_ruby.rbpsych/visitors/visitor.rbpsych/visitors/yaml_tree.rbpsych/y.rb

Gem::Specification.new do |s|
  s.name = "psych".freeze
  s.version = "3.3.0"

  s.required_rubygems_version = Gem::Requirement.new(">= 0".freeze) if s.respond_to? :required_rubygems_version=
  s.require_paths = ["lib".freeze]
  s.authors = ["Aaron Patterson".freeze, "SHIBATA Hiroshi".freeze, "Charles Oliver Nutter".freeze]
  s.date = "2021-07-08"
  s.description = "Psych is a YAML parser and emitter. Psych leverages libyaml[https://pyyaml.org/wiki/LibYAML]\nfor its YAML parsing and emitting capabilities. In addition to wrapping libyaml,\nPsych also knows how to serialize and de-serialize most Ruby objects to and from the YAML format.\n".freeze
  s.email = ["aaron@tenderlovemaking.com".freeze, "hsbt@ruby-lang.org".freeze, "headius@headius.com".freeze]
  s.extensions = ["ext/psych/extconf.rb".freeze]
  s.extra_rdoc_files = ["README.md".freeze]
  s.files = ["README.md".freeze, "ext/psych/extconf.rb".freeze, "psych.rb".freeze, "psych.so".freeze, "psych/class_loader.rb".freeze, "psych/coder.rb".freeze, "psych/core_ext.rb".freeze, "psych/exception.rb".freeze, "psych/handler.rb".freeze, "psych/handlers/document_stream.rb".freeze, "psych/handlers/recorder.rb".freeze, "psych/json/ruby_events.rb".freeze, "psych/json/stream.rb".freeze, "psych/json/tree_builder.rb".freeze, "psych/json/yaml_events.rb".freeze, "psych/nodes.rb".freeze, "psych/nodes/alias.rb".freeze, "psych/nodes/document.rb".freeze, "psych/nodes/mapping.rb".freeze, "psych/nodes/node.rb".freeze, "psych/nodes/scalar.rb".freeze, "psych/nodes/sequence.rb".freeze, "psych/nodes/stream.rb".freeze, "psych/omap.rb".freeze, "psych/parser.rb".freeze, "psych/scalar_scanner.rb".freeze, "psych/set.rb".freeze, "psych/stream.rb".freeze, "psych/streaming.rb".freeze, "psych/syntax_error.rb".freeze, "psych/tree_builder.rb".freeze, "psych/versions.rb".freeze, "psych/visitors.rb".freeze, "psych/visitors/depth_first.rb".freeze, "psych/visitors/emitter.rb".freeze, "psych/visitors/json_tree.rb".freeze, "psych/visitors/to_ruby.rb".freeze, "psych/visitors/visitor.rb".freeze, "psych/visitors/yaml_tree.rb".freeze, "psych/y.rb".freeze]
  s.homepage = "https://github.com/ruby/psych".freeze
  s.licenses = ["MIT".freeze]
  s.rdoc_options = ["--main".freeze, "README.md".freeze]
  s.required_ruby_version = Gem::Requirement.new(">= 2.4.0".freeze)
  s.rubygems_version = "3.3.0.dev".freeze
  s.summary = "Psych is a YAML parser and emitter".freeze
end
```
That way we can use the existing mechanism to avoid evaulating gemspecs if stub lines already include the information we need.

When rubygems is updated, all default gemspecs will be regenerated to include this line. When rubygems is downgraded, gemspecs are left there without that line, but I don't think that will cause any issues since the comment will be ignored by older rubygems.

The results I'm getting with this optimization are pretty unstable, but they show a consistent speedup. Compared to current master branch, rubygems boots ~20% faster, and the overhead of rubygems over ruby boot time goes from ~75% slower to ~50% slower.

```
$ hyperfine --prepare 'git checkout speedup && rake install' --command-name optimization 'ruby -e1' --prepare 'git checkout master && rake install' --command-name current 'ruby -e1'
Benchmark #1: optimization
  Time (mean ± σ):      71.8 ms ±   5.7 ms    [User: 59.0 ms, System: 15.0 ms]
  Range (min … max):    66.6 ms …  85.3 ms    10 runs
 
Benchmark #2: current
  Time (mean ± σ):      84.4 ms ±   4.2 ms    [User: 73.5 ms, System: 13.2 ms]
  Range (min … max):    79.2 ms …  90.7 ms    10 runs
 
Summary
  'optimization' ran
    1.18 ± 0.11 times faster than 'current'
```

```
$ hyperfine --prepare 'git checkout speedup && rake install' --command-name optimization 'ruby -e1' --prepare 'git checkout master && rake install' --command-name current 'ruby -e1' --prepare ''  --command-name without-rubygems 'ruby --disable-gems -e1'
Benchmark #1: optimization
  Time (mean ± σ):      67.8 ms ±   1.1 ms    [User: 59.9 ms, System: 10.0 ms]
  Range (min … max):    66.9 ms …  70.5 ms    10 runs
 
Benchmark #2: current
  Time (mean ± σ):      80.5 ms ±   0.7 ms    [User: 71.8 ms, System: 10.8 ms]
  Range (min … max):    79.7 ms …  81.7 ms    10 runs
 
Benchmark #3: without-rubygems
  Time (mean ± σ):      46.6 ms ±  10.4 ms    [User: 32.1 ms, System: 17.1 ms]
  Range (min … max):    33.0 ms …  78.1 ms    77 runs
 
Summary
  'without-rubygems' ran
    1.46 ± 0.33 times faster than 'optimization'
    1.73 ± 0.39 times faster than 'current'
```

## Make sure the following tasks are checked

- [x] Describe the problem / feature
- [ ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [x] Write code to solve the problem
- [x] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)