@deivid-rodriguez 2020-12-28
I'm working on more boot performance improvements and now I have the following performance locally:

```
Benchmark #1: ruby -e1
  Time (mean ± σ):      56.2 ms ±   1.9 ms    [User: 46.3 ms, System: 12.2 ms]
  Range (min … max):    53.9 ms …  62.4 ms    52 runs
 
Benchmark #2: ruby --disable-gems -e1
  Time (mean ± σ):      43.8 ms ±   8.5 ms    [User: 31.4 ms, System: 15.2 ms]
  Range (min … max):    34.3 ms …  55.3 ms    84 runs
 
Summary
  'ruby --disable-gems -e1' ran
    1.28 ± 0.25 times faster than 'ruby -e1'
```

Will try to get this ready soon.