@deivid-rodriguez 2021-07-12
I updated this PR. The results I posted in the previous message were with further optimizations, but I think I'll propose this first and then the other optimizations separately.