@deivid-rodriguez 2021-07-14
Also added some code locally to compare this approach with the one proposed at https://github.com/rubygems/rubygems/issues/3799 of caching default marshalled specs, and it seems that this approach is consistently faster:

```
$ hyperfine --warmup 1 --command-name current 'ruby -e1' --command-name stub_default_spec_files 'STUB_DEFAULT_SPEC_FILES=true ruby -e1' --command-name 'cache_default_specs' 'CACHE_DEFAULT_SPECS=true ruby -e1'
Benchmark #1: current
  Time (mean ± σ):      79.0 ms ±   1.5 ms    [User: 68.1 ms, System: 13.3 ms]
  Range (min … max):    77.5 ms …  85.1 ms    37 runs
 
Benchmark #2: stub_default_spec_files
  Time (mean ± σ):      66.3 ms ±   1.5 ms    [User: 56.5 ms, System: 12.1 ms]
  Range (min … max):    64.5 ms …  71.6 ms    43 runs
 
Benchmark #3: cache_default_specs
  Time (mean ± σ):      75.6 ms ±   1.7 ms    [User: 64.8 ms, System: 13.1 ms]
  Range (min … max):    73.2 ms …  79.2 ms    38 runs
 
Summary
  'stub_default_spec_files' ran
    1.14 ± 0.04 times faster than 'cache_default_specs'
    1.19 ± 0.03 times faster than 'current'
```