@eregon 2022-02-03
The numbers for `ruby --disable-gems` look very high to me, maybe you're using rbenv or RVM?
For startup measurement it's best to ensure the real ruby binary is already in PATH, without any wrapper before.
Or potentially just pass a full path to the real ruby binary.