@deivid-rodriguez 2022-08-06
I rebased this and re-measured with my current Ruby installation which does not use shims:

```
$ hyperfine -N --command-name cache-default-marshaled-specs --prepare "./tmp/prepare.sh speedup-alt" "ruby -e1"  --command-name this-pr --prepare "./tmp/prepare.sh speedup" "ruby -e1" --command-name master --prepare "./tmp/prepare.sh master" "ruby -e1" --warmup 5
Benchmark 1: cache-default-marshaled-specs
  Time (mean ± σ):      29.5 ms ±   0.1 ms    [User: 23.2 ms, System: 5.5 ms]
  Range (min … max):    29.3 ms …  29.7 ms    10 runs
 
Benchmark 2: this-pr
  Time (mean ± σ):      27.4 ms ±   0.4 ms    [User: 20.8 ms, System: 5.8 ms]
  Range (min … max):    27.1 ms …  28.4 ms    10 runs
 
Benchmark 3: master
  Time (mean ± σ):      32.1 ms ±   0.2 ms    [User: 25.3 ms, System: 5.9 ms]
  Range (min … max):    31.9 ms …  32.5 ms    10 runs
 
Summary
  'this-pr' ran
    1.08 ± 0.01 times faster than 'cache-default-marshaled-specs'
    1.17 ± 0.02 times faster than 'master'
```