@eregon 2022-08-06
I'm a little bit worried to store this information in the `.gemspec` for two reasons:
* It can make the gemspec a lot bigger (a gem can have many files), and so if we actually need to load it fully we'll take longer to read it
* It makes the gemspec unreadable in most editors/CLI tools due to the `\0` separators, which is unfortunate when debugging/investigating (already an issue with `stub` lines).

Would it make sense to have a separate file for this information? That would also enable storing the information in a custom format, and not need to buffer a large part of the file just to find the line endings, which would likely mean additional speedups.