@deivid-rodriguez 2022-08-06
> It can make the gemspec a lot bigger (a gem can have many files), and so if we actually need to load it fully we'll take longer to read it

Default gems don't usually have many files, and we only do this for default gems. Do you really think this can be a noticeable problem?

> It makes the gemspec unreadable in most editors/CLI tools due to the \0 separators, which is unfortunate when debugging/investigating (already an issue with stub lines).

As you said, this is an existing issue, and it seems acceptable to me. We can consider changing it, there's an open issue for that, but only a single person reported this as an issue in like 15 years.

> Would it make sense to have a separate file for this information? That would also enable storing the information in a binary format, and not need to buffer a large part of the file just to find the line endings, which would likely mean additional speedups.

I guess, but I think this is an improvement already and the downsides seem quite weak.