@eregon 2022-08-06
> Default gems don't usually have many files, and we only do this for default gems. Do you really think this can be a noticeable problem?

Yes, otherwise I wouldn't mention it. It would basically double the size of every default gem `.gemspec` file.
(talking about file size, the `.freeze` everywhere look not nice and are a waste of file size, `# frozen_string_literal: true` seems much better)

Since `s.files =` seems by far the largest field of a gemspec, it is a bit surprising that this approach gains something. Maybe there is some non-trivial costs for other fields being assigned? Otherwise we're just measuring Ruby parser speed vs read with buffer, split by lines, split by \0, at the cost of storing the information twice.

```
> Gem::Specification.each.select(&:default_gem?).map { [_1.files.size, _1.name] }.sort
...
 [8, "optparse"],
 [11, "net-http"],
 [12, "fiddle"],
 [14, "csv"],
 [14, "digest"],
 [14, "drb"],
 [15, "uri"],
 [17, "reline"],
 [18, "did_you_mean"],
 [19, "openssl"],
 [26, "json"],
 [26, "racc"],
 [40, "psych"],
 [46, "irb"],
 [124, "rdoc"],
 [211, "bundler"]]
```

> but only a single person reported this as an issue in like 15 years.

Yet multiple people seem annoyed by it, IMHO worth fixing. I didn't report it until now but I've been annoyed by that several times (when debugging stuff related to default gems).
I'm probably more exposed to it than the average Ruby user though.